+++ axelor-account/src/main/java/com/axelor/apps/account/db/repo/PaymentVoucherManagementRepository.java
@@ -20,6 +20,7 @@ package com.axelor.apps.account.db.repo;
 import com.axelor.apps.account.db.PaymentVoucher;
 import com.axelor.apps.account.service.payment.paymentvoucher.PaymentVoucherSequenceService;
 import com.axelor.apps.base.service.app.AppBaseService;
+import com.axelor.exception.service.TraceBackService;
 import com.axelor.inject.Beans;
 import javax.persistence.PersistenceException;
 
@@ -59,6 +60,7 @@ public class PaymentVoucherManagementRepository extends PaymentVoucherRepository
 
       return super.save(paymentVoucher);
     } catch (Exception e) {
+      TraceBackService.trace(e);
       throw new PersistenceException(e.getLocalizedMessage());
     }
   }