+++ axelor-contract/src/main/java/com/axelor/apps/contract/service/ContractServiceImpl.java
@@ -18,10 +18,12 @@
 package com.axelor.apps.contract.service;
 
 import com.axelor.apps.account.db.Account;
+import com.axelor.apps.account.db.AnalyticMoveLine;
 import com.axelor.apps.account.db.FiscalPosition;
 import com.axelor.apps.account.db.Invoice;
 import com.axelor.apps.account.db.InvoiceLine;
 import com.axelor.apps.account.db.TaxLine;
+import com.axelor.apps.account.db.repo.AnalyticMoveLineRepository;
 import com.axelor.apps.account.db.repo.InvoiceLineRepository;
 import com.axelor.apps.account.db.repo.InvoiceRepository;
 import com.axelor.apps.account.service.FiscalPositionAccountService;
@@ -441,6 +443,7 @@ public class ContractServiceImpl extends ContractRepository implements ContractS
 
       for (ContractLine line : lines) {
         ContractLine tmp = contractLineRepo.copy(line, false);
+        tmp.setAnalyticMoveLineList(line.getAnalyticMoveLineList());
         tmp.setQty(tmp.getQty().multiply(ratio).setScale(QTY_SCALE, RoundingMode.HALF_UP));
         tmp = this.contractLineService.computeTotal(tmp);
         generate(invoice, tmp);
@@ -577,9 +580,32 @@ public class ContractServiceImpl extends ContractRepository implements ContractS
     invoiceLine.setTaxLine(taxLine);
     invoiceLine.setAccount(replacedAccount);
 
+    if (line.getAnalyticDistributionTemplate() != null) {
+      invoiceLine.setAnalyticDistributionTemplate(line.getAnalyticDistributionTemplate());
+      this.copyAnalyticMoveLines(line.getAnalyticMoveLineList(), invoiceLine);
+    }
+
     invoice.addInvoiceLineListItem(invoiceLine);
 
     return Beans.get(InvoiceLineRepository.class).save(invoiceLine);
+  }
+
+  public void copyAnalyticMoveLines(
+      List<AnalyticMoveLine> originalAnalyticMoveLineList, InvoiceLine invoiceLine) {
+    if (originalAnalyticMoveLineList == null) {
+      return;
+    }
+
+    AnalyticMoveLineRepository analyticMoveLineRepo = Beans.get(AnalyticMoveLineRepository.class);
+
+    for (AnalyticMoveLine originalAnalyticMoveLine : originalAnalyticMoveLineList) {
+      AnalyticMoveLine analyticMoveLine =
+          analyticMoveLineRepo.copy(originalAnalyticMoveLine, false);
+
+      analyticMoveLine.setTypeSelect(AnalyticMoveLineRepository.STATUS_FORECAST_INVOICE);
+      analyticMoveLine.setContractLine(null);
+      invoiceLine.addAnalyticMoveLineListItem(analyticMoveLine);
+    }
   }
 
   @Override