+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/StockMoveServiceSupplychainImpl.java
@@ -27,19 +27,21 @@ import com.axelor.apps.sale.db.SaleOrder;
 import com.axelor.apps.sale.db.SaleOrderLine;
 import com.axelor.apps.sale.db.repo.SaleOrderLineRepository;
 import com.axelor.apps.sale.db.repo.SaleOrderRepository;
-import com.axelor.apps.stock.db.StockLocationLine;
 import com.axelor.apps.stock.db.StockMove;
 import com.axelor.apps.stock.db.StockMoveLine;
 import com.axelor.apps.stock.db.repo.StockMoveLineRepository;
 import com.axelor.apps.stock.db.repo.StockMoveRepository;
 import com.axelor.apps.stock.service.PartnerProductQualityRatingService;
-import com.axelor.apps.stock.service.StockLocationLineService;
 import com.axelor.apps.stock.service.StockMoveLineService;
 import com.axelor.apps.stock.service.StockMoveServiceImpl;
 import com.axelor.apps.stock.service.StockMoveToolService;
+import com.axelor.apps.supplychain.exception.IExceptionMessage;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
 import com.axelor.exception.AxelorException;
+import com.axelor.exception.db.repo.TraceBackRepository;
+import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
+import com.google.common.base.Strings;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
 import java.lang.invoke.MethodHandles;
@@ -47,6 +49,7 @@ import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.StringJoiner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -60,6 +63,8 @@ public class StockMoveServiceSupplychainImpl extends StockMoveServiceImpl
   protected SaleOrderRepository saleOrderRepo;
   protected UnitConversionService unitConversionService;
   protected ReservedQtyService reservedQtyService;
+
+  @Inject private StockMoveLineServiceSupplychain stockMoveLineServiceSupplychain;
 
   @Inject
   public StockMoveServiceSupplychainImpl(
@@ -381,14 +386,29 @@ public class StockMoveServiceSupplychainImpl extends StockMoveServiceImpl
   }
 
   @Override
-  public BigDecimal getAvailableStock(StockMove stockMove, StockMoveLine stockMoveLine) {
-    StockLocationLine stockLocationLine =
-        Beans.get(StockLocationLineService.class)
-            .getStockLocationLine(stockMove.getFromStockLocation(), stockMoveLine.getProduct());
-
-    if (stockLocationLine == null) {
-      return BigDecimal.ZERO;
+  public void verifyProductStock(StockMove stockMove) throws AxelorException {
+    AppSupplychain appSupplychain = appSupplyChainService.getAppSupplychain();
+    if (stockMove.getAvailabilityRequest()
+        && stockMove.getStockMoveLineList() != null
+        && appSupplychain.getIsVerifyProductStock()
+        && stockMove.getFromStockLocation() != null) {
+      StringJoiner notAvailableProducts = new StringJoiner(",");
+      int counter = 1;
+      for (StockMoveLine stockMoveLine : stockMove.getStockMoveLineList()) {
+        boolean isAvailableProduct =
+            stockMoveLineServiceSupplychain.isAvailableProduct(stockMove, stockMoveLine);
+        if (!isAvailableProduct && counter <= 10) {
+          notAvailableProducts.add(stockMoveLine.getProduct().getFullName());
+          counter++;
+        }
+      }
+      if (!Strings.isNullOrEmpty(notAvailableProducts.toString())) {
+        throw new AxelorException(
+            TraceBackRepository.CATEGORY_CONFIGURATION_ERROR,
+            String.format(
+                I18n.get(IExceptionMessage.STOCK_MOVE_VERIFY_PRODUCT_STOCK_ERROR),
+                notAvailableProducts.toString()));
+      }
     }
-    return stockLocationLine.getCurrentQty().subtract(stockLocationLine.getReservedQty());
   }
 }